Fix basic-export: write zip files instead of tsv
